[TextRenderer::compute_text_radii()] Cast properly

+++ src/details/TextRenderer.cpp
@@ -57,7 +57,8 @@ namespace TextRendererU {
 
 Radii compute_text_radii(const std::u16string& text, float font_size)
 {
-    return font_size * glm::vec2{text.length(), 1.f};
+    return font_size * glm::vec2{static_cast<float>(text.length()),
+                                 1.f};
 }
 
 } // namespace TextRendererU
